+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ### Fixed
 - Corrected `New > Zephir File` dialog to use project's directory as a namespace
 
+### Removed
+- Drop support of IDE's versions below `191`
+
 ## [0.3.6] - 2020-05-18
 ### Changed
 - Updated file icons